4.1.6 fix mistakes, separate util functions, add function for create date and timing
