+++ src/components/film-card.js
@@ -1,4 +1,5 @@
-import {createElement} from '../utils.js';
+import {createElement} from '../utils/utils.js';
+import {getRatingForInsertion, getdurationInHours} from '../utils/film-utils.js';
 
 const createFilmButtonMarkup = (filmButton, isActive) => {
   const [type, name] = filmButton;
@@ -10,17 +11,16 @@ const createFilmButtonMarkup = (filmButton, isActive) => {
 const createFilmCardTemplate = (movie) => {
   const {title, rating, date, duration, genres, poster, description, onWatchlist, isWatched, onFavorite, comments} = movie;
 
-  const ratingForInsertion = rating % 1 ? rating : rating + `.0`;
-  const year = date.getDate();
+  const year = date.getFullYear();
   const descriptionForInsertion = description.length > 140 ? (description.substring(0, 139) + `...`) : description;
 
   return (
     `<article class="film-card">
       <h3 class="film-card__title">${title}</h3>
-      <p class="film-card__rating">${ratingForInsertion}</p>
+      <p class="film-card__rating">${getRatingForInsertion(rating)}</p>
       <p class="film-card__info">
         <span class="film-card__year">${year}</span>
-        <span class="film-card__duration">${duration}</span>
+        <span class="film-card__duration">${getdurationInHours(duration)}</span>
         <span class="film-card__genre">${genres[0]}</span>
       </p>
       <img src="./images/posters/${poster}" alt="film poster ${title}" class="film-card__poster">
