+++ src/components/film-container.js
@@ -1,4 +1,4 @@
-import {createElement} from '../utils.js';
+import {createElement} from '../utils/utils.js';
 
 const createFilmContainerTemplate = () => {
   return (
