+++ src/components/user-profile.js
@@ -1,4 +1,4 @@
-import {createElement} from '../utils.js';
+import {createElement} from '../utils/utils.js';
 
 const createUserProfileTemplate = () => {
   return (
