+++ src/mock/film.js
@@ -1,4 +1,5 @@
-import {getRandomInteger, getRandomRational, getRandomArrayItem, getArrayRandomItems, getRandomDate} from "../utils.js";
+import {getRandomInteger, getRandomRational, getRandomArrayItem, getArrayRandomItems} from '../utils/utils.js';
+import {getRandomDate} from '../utils/film-utils.js';
 
 const FILM_TITLES = [
   {
